Removed the std::shared_ptr semantics - it's cumbersome.
The whole idea was really an experiment to see if that was easier to use.
Turns out it doesn't look so nice from the ownership perspective, because
users will still expect each object to behave independently. Instead, new
objects will copy the objects as necessary.

The alternative would be to add logic so that whenever a Quadrature is resized,
we would also resize any dependent objects (and recalculate them). This is more
logic than I would like to see in the implementation right now.